refactor(topics): drop dead special_category column and unreachable branch
Second review pass follow-ups:

- Remove the forensic_candidates.special_category column, its CHECK, and the
  per-insert extraction: no query ever filtered or sorted on it and the flag
  already lives in fields_json. Correct the insert comment to describe only the
  topic_label projection that the query actually uses.
- Delete the unreachable mutual-exclusivity branch in selectEmittedLabels:
  `unclassified` is excluded from the anchors, so it is never scored and can
  never appear in the ranked set — the filter enforced a false invariant.
- Extract the History topic-classification pass (classifyHistoryTopics +
  topicInputsFromArtifact + TopicCandidateSummary) into
  history-topic-classification.ts, and collapse the five near-identical summary
  object literals behind disabled/unavailable/complete helpers.
