+++ core/src/case-findings.ts
@@ -409,8 +409,9 @@ export function initializeFindingSchema(database: DatabaseSync): void {
       ),
       rank INTEGER NOT NULL CHECK (rank > 0),
       supporting_count INTEGER NOT NULL CHECK (supporting_count > 0),
+      -- topic-specific projection of fields.topicLabel, kept as a column so the
+      -- Candidate query can filter and sort by label without parsing JSON.
       topic_label TEXT,
-      special_category INTEGER NOT NULL DEFAULT 0 CHECK (special_category IN (0, 1)),
       provenance_json TEXT NOT NULL,
       fields_json TEXT NOT NULL,
       search_text TEXT NOT NULL
@@ -1017,19 +1018,14 @@ function insertCandidate(
   row: PersistedCandidate,
 ): void {
   const candidate = row.candidate;
-  // Denormalise the topic label and special-category flag into columns so the
-  // Candidate query can filter and sort on them without parsing JSON per row.
-  // They are a projection of `fields`, never a second source of truth.
+  // Denormalise the topic label into a column so the Candidate query can filter
+  // and sort by it without parsing JSON per row. It is a projection of
+  // `fields.topicLabel`, never a second source of truth.
   const labelField = candidate.fields.topicLabel;
   const topicLabel =
     labelField !== undefined && labelField.state === "value"
       ? String(labelField.value)
       : null;
-  const scField = candidate.fields.specialCategory;
-  const specialCategory =
-    scField !== undefined && scField.state === "value" && scField.value === true
-      ? 1
-      : 0;
   statement.run(
     artifactResultId,
     runId,
@@ -1040,7 +1036,6 @@ function insertCandidate(
     candidate.rank,
     candidate.count,
     topicLabel,
-    specialCategory,
     JSON.stringify(candidate.provenance),
     JSON.stringify(candidate.fields),
     row.searchText,
@@ -1368,9 +1363,8 @@ export function storeHistoryAnalysis(
         `INSERT INTO forensic_candidates
            (artifact_result_id, run_id, record_type, candidate_kind,
             profile_path, commit_state, rank, supporting_count,
-            topic_label, special_category, provenance_json, fields_json,
-            search_text)
-         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
+            topic_label, provenance_json, fields_json, search_text)
+         VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`,
       );
       const deactivatePrevious = database.prepare(
         `UPDATE history_artifact_results
